pyproject.toml
[project]
name = "backend-image-analyser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.124.4",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-jose[cryptography]>=3.5.0",
    "python-json-logger>=4.0.0",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "redis[hiredis]>=7.1.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.45",
    "tenacity>=9.1.2",
    "uvicorn>=0.38.0",
]

[tool.setuptools]
packages = ["app"]
